+++ src/core/page/destroy.js
@@ -1,6 +1,8 @@
 
 'use strict';
 
+var vars = require('../../data/vars');
+
 module.exports = (self) => {
     // call all fragments's __orchids__destroy
     Object.keys(self.__orchids__fragmentsInstances).forEach((id) => {
@@ -13,7 +15,7 @@ module.exports = (self) => {
         // has animation
         setTimeout(() => {
             self.el.remove()
-        }, 500)
+        }, vars.animateTime)
     }
     else
     // no animation