+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - make `container` instead of `rootContainer` of `orchids.init`
 - make `onResult` instead of `onPageResult, onDialogResult`
+- add `beforeCreate, afterCreate` to `page, dialog`
 
 ## 0.1.1